refactor: move regularisation routines to `_regularize`
refactor: move regularisation routines to `_regularize`
