+++ src/awkward/contents/recordarray.py
@@ -10,6 +10,7 @@ from awkward._behavior import find_record_reducer
 from awkward._nplikes.numpy import Numpy
 from awkward._nplikes.numpylike import IndexType, NumpyMetadata
 from awkward._nplikes.shape import unknown_length
+from awkward._regularize import is_integer
 from awkward._util import unset
 from awkward.contents.content import Content
 from awkward.forms.form import _type_parameters_equal
@@ -211,7 +212,7 @@ class RecordArray(Content):
                         )
                     )
 
-            if not (ak._util.is_integer(length) and length >= 0):
+            if not (is_integer(length) and length >= 0):
                 raise ak._errors.wrap_error(
                     TypeError(
                         "{} 'length' must be a non-negative integer or None, not {}".format(
