+++ src/awkward/operations/ak_combinations.py
@@ -1,7 +1,7 @@
 # BSD 3-Clause License; see https://github.com/scikit-hep/awkward-1.0/blob/main/LICENSE
-
 import awkward as ak
 from awkward._nplikes.numpylike import NumpyMetadata
+from awkward._regularize import regularize_axis
 
 np = NumpyMetadata.instance()
 
@@ -203,7 +203,7 @@ def combinations(
 def _impl(
     array, n, replacement, axis, fields, parameters, with_name, highlevel, behavior
 ):
-    axis = ak._util.regularize_axis(axis)
+    axis = regularize_axis(axis)
     if parameters is None:
         parameters = {}
     else:
