+++ src/awkward/operations/ak_linear_fit.py
@@ -4,6 +4,7 @@ from awkward._backends import backend_of
 from awkward._behavior import behavior_of
 from awkward._nplikes import ufuncs
 from awkward._nplikes.numpylike import NumpyMetadata
+from awkward._regularize import regularize_axis
 from awkward._util import unset
 
 np = NumpyMetadata.instance()
@@ -88,7 +89,7 @@ def linear_fit(
 
 
 def _impl(x, y, weight, axis, keepdims, mask_identity):
-    axis = ak._util.regularize_axis(axis)
+    axis = regularize_axis(axis)
     behavior = behavior_of(x, y, weight)
     x = ak.highlevel.Array(
         ak.operations.to_layout(x, allow_record=False, allow_other=False),
