+++ src/awkward/operations/ak_min.py
@@ -3,6 +3,7 @@ import awkward as ak
 from awkward._behavior import behavior_of
 from awkward._connect.numpy import unsupported
 from awkward._nplikes.numpylike import NumpyMetadata
+from awkward._regularize import regularize_axis
 from awkward._util import unset
 
 np = NumpyMetadata.instance()
@@ -166,7 +167,7 @@ def nanmin(
 
 
 def _impl(array, axis, keepdims, initial, mask_identity, highlevel, behavior):
-    axis = ak._util.regularize_axis(axis)
+    axis = regularize_axis(axis)
     layout = ak.operations.to_layout(array, allow_record=False, allow_other=False)
     behavior = behavior_of(array, behavior=behavior)
     reducer = ak._reducers.Min(initial)
