+++ src/awkward/operations/ak_pad_none.py
@@ -1,7 +1,7 @@
 # BSD 3-Clause License; see https://github.com/scikit-hep/awkward-1.0/blob/main/LICENSE
-
 import awkward as ak
 from awkward._nplikes.numpylike import NumpyMetadata
+from awkward._regularize import regularize_axis
 
 np = NumpyMetadata.instance()
 
@@ -108,7 +108,7 @@ def pad_none(array, target, axis=1, *, clip=False, highlevel=True, behavior=None
 
 
 def _impl(array, target, axis, clip, highlevel, behavior):
-    axis = ak._util.regularize_axis(axis)
+    axis = regularize_axis(axis)
     layout = ak.operations.to_layout(array, allow_record=False, allow_other=False)
     out = ak._do.pad_none(layout, target, axis, clip=clip)
