+++ src/awkward/operations/ak_sort.py
@@ -1,8 +1,8 @@
 # BSD 3-Clause License; see https://github.com/scikit-hep/awkward-1.0/blob/main/LICENSE
-
 import awkward as ak
 from awkward._connect.numpy import unsupported
 from awkward._nplikes.numpylike import NumpyMetadata
+from awkward._regularize import regularize_axis
 
 np = NumpyMetadata.instance()
 
@@ -47,7 +47,7 @@ def sort(array, axis=-1, *, ascending=True, stable=True, highlevel=True, behavio
 
 
 def _impl(array, axis, ascending, stable, highlevel, behavior):
-    axis = ak._util.regularize_axis(axis)
+    axis = regularize_axis(axis)
     layout = ak.operations.to_layout(array, allow_record=False, allow_other=False)
     out = ak._do.sort(layout, axis, ascending, stable)
     return ak._util.wrap_layout(out, behavior, highlevel, like=array)
