+++ src/awkward/operations/ak_std.py
@@ -4,6 +4,7 @@ from awkward._behavior import behavior_of
 from awkward._connect.numpy import unsupported
 from awkward._nplikes import ufuncs
 from awkward._nplikes.numpylike import NumpyMetadata
+from awkward._regularize import regularize_axis
 from awkward._util import unset
 
 np = NumpyMetadata.instance()
@@ -159,7 +160,7 @@ def nanstd(
 
 
 def _impl(x, weight, ddof, axis, keepdims, mask_identity):
-    axis = ak._util.regularize_axis(axis)
+    axis = regularize_axis(axis)
     behavior = behavior_of(x, weight)
     x = ak.highlevel.Array(
         ak.operations.to_layout(x, allow_record=False, allow_other=False),
