+++ src/awkward/operations/ak_var.py
@@ -3,6 +3,7 @@ import awkward as ak
 from awkward._behavior import behavior_of
 from awkward._connect.numpy import unsupported
 from awkward._nplikes.numpylike import NumpyMetadata
+from awkward._regularize import regularize_axis
 from awkward._util import unset
 
 np = NumpyMetadata.instance()
@@ -164,7 +165,7 @@ def nanvar(
 
 
 def _impl(x, weight, ddof, axis, keepdims, mask_identity):
-    axis = ak._util.regularize_axis(axis)
+    axis = regularize_axis(axis)
     behavior = behavior_of(x, weight)
     x = ak.highlevel.Array(
         ak.operations.to_layout(x, allow_record=False, allow_other=False),
