+++ src/awkward/record.py
@@ -8,6 +8,7 @@ import awkward as ak
 from awkward._behavior import get_record_class
 from awkward._nplikes.numpylike import NumpyMetadata
 from awkward._nplikes.shape import unknown_length
+from awkward._regularize import is_integer
 from awkward._util import unset
 from awkward.contents.content import Content
 from awkward.typing import Self
@@ -35,7 +36,7 @@ class Record:
             raise ak._errors.wrap_error(
                 TypeError(f"Record 'array' must be a RecordArray, not {array!r}")
             )
-        if not ak._util.is_integer(at):
+        if not is_integer(at):
             raise ak._errors.wrap_error(
                 TypeError(f"Record 'at' must be an integer, not {array!r}")
             )
@@ -137,7 +138,7 @@ class Record:
             return self._getitem(where)
 
     def _getitem(self, where):
-        if ak._util.is_integer(where):
+        if is_integer(where):
             raise ak._errors.wrap_error(
                 IndexError("scalar Record cannot be sliced by an integer")
             )
